machine/usb/hid/joystick: move joystick under HID as it belongs and also remove duplicate code
