+++ src/machine/usb/hid/joystick/state.go
